refactor models Ids

+++ src/main/java/com/cailou/esigAtividadeTecnica/repository/TokenRepository.java
@@ -3,5 +3,7 @@ package com.cailou.esigAtividadeTecnica.repository;
 import com.cailou.esigAtividadeTecnica.model.TokenModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TokenRepository extends JpaRepository<TokenModel, Long> {
+import java.math.BigInteger;
+
+public interface TokenRepository extends JpaRepository<TokenModel, BigInteger> {
 }
